pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medmnist-mini-bench"
version = "0.1.0"
description = "Minimal, reproducible MedMNIST baselines with 2D/3D CNNs and one-command demos."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [{name = "Xiaoliang"}]
dependencies = [
  "medmnist>=3.0.1",
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "numpy",
  "pandas",
  "scikit-learn",
  "matplotlib",
  "pyyaml"
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bench = ["*.yaml"]
